feat: 인원 탈주시에만 검색창이 뜨도록 설정
