+++ client/src/components/Modal/SetActionModal/AddMemberActionListModalContent/AddMemberActionListModalContent.tsx
@@ -5,11 +5,11 @@ import {FixedButton, LabelGroupInput} from 'haengdong-design';
 import {useStepList} from '@hooks/useStepList/useStepList';
 import validateMemberName from '@utils/validate/validateMemberName';
 
-import useSearchInMemberList from '@hooks/useSearchInMemberList';
 import useDynamicInput from '@hooks/useDynamicInput';
 
 import style from './AddMemberActionListModalContent.style';
-import Search from './Search/Search';
+import InMember from './InMember';
+import OutMember from './OutMember';
 
 interface AddMemberActionListModalContentProps {
   inOutAction: MemberType;
@@ -17,17 +17,8 @@ interface AddMemberActionListModalContentProps {
 }
 
 const AddMemberActionListModalContent = ({inOutAction, setIsOpenBottomSheet}: AddMemberActionListModalContentProps) => {
-  const {
-    inputList,
-    inputRefList,
-    handleInputChange,
-    deleteEmptyInputElementOnBlur,
-    getFilledInputList,
-    errorMessage,
-    canSubmit,
-    focusNextInputOnEnter,
-    setInputValueTargetIndex,
-  } = useDynamicInput(validateMemberName);
+  const dynamicProps = useDynamicInput(validateMemberName);
+  const {inputList, getFilledInputList, errorMessage, canSubmit} = dynamicProps;
 
   const {updateMemberList} = useStepList();
 
@@ -36,39 +27,12 @@ const AddMemberActionListModalContent = ({inOutAction, setIsOpenBottomSheet}: Ad
     setIsOpenBottomSheet(false);
   };
 
-  const {filteredInMemberList, currentInputIndex, handleCurrentInputIndex, searchCurrentInMember, chooseMember} =
-    useSearchInMemberList(setInputValueTargetIndex);
-
-  const validationAndSearchOnChange = (inputIndex: number, event: React.ChangeEvent<HTMLInputElement>) => {
-    handleCurrentInputIndex(inputIndex);
-    handleInputChange(inputIndex, event);
-    searchCurrentInMember(event);
-  };
-
   return (
     <div css={style.container}>
       <div css={style.inputGroup}>
         {/* TODO: (@soha) Search로 변경하기 */}
         <LabelGroupInput labelText="이름" errorText={errorMessage}>
-          {inputList.map(({value, index}) => (
-            <Search
-              isShow={currentInputIndex === index}
-              searchTerms={filteredInMemberList}
-              onTermClick={(term: string) => chooseMember(currentInputIndex, term)}
-            >
-              <LabelGroupInput.Element
-                key={index}
-                elementKey={`${index}`}
-                type="text"
-                value={value}
-                ref={el => (inputRefList.current[index] = el)}
-                onChange={e => validationAndSearchOnChange(index, e)}
-                onBlur={() => deleteEmptyInputElementOnBlur()}
-                onKeyDown={e => focusNextInputOnEnter(e, index)}
-                placeholder="이름"
-              />
-            </Search>
-          ))}
+          {inOutAction === 'IN' ? <InMember dynamicProps={dynamicProps} /> : <OutMember dynamicProps={dynamicProps} />}
         </LabelGroupInput>
       </div>
       <FixedButton
